python/requirements.txt
asyncio==3.4.3
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
future==0.18.2
gkeepapi==0.13.4
gpsoauth==0.4.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycryptodomex==3.9.9
pyserial==3.5
requests==2.23.0
six==1.15.0
sockets==1.0.0
urllib3==1.25.11
websockets==8.0.2
Werkzeug==1.0.1
